--- app/src/main/java/de/coop/tgvertretung/Statics.java
@@ -1,5 +1,0 @@
-package de.coop.tgvertretung;
-
-public class Statics {
-    public static MainActivity mainActivity = null;
-}
