+++ app/src/main/java/de/coop/tgvertretung/utils/ObjectSerializer.java
@@ -1,4 +1,4 @@
-package de.coop.tgvertretung;
+package de.coop.tgvertretung.utils;
 
 import android.util.Base64;
