+++ app/src/main/java/de/coop/tgvertretung/utils/Statics.java
@@ -1,0 +1,7 @@
+package de.coop.tgvertretung.utils;
+
+import de.coop.tgvertretung.activity.MainActivity;
+
+public class Statics {
+    public static MainActivity mainActivity = null;
+}
